pyproject.toml
[project]
name = "bloqade-qourier" # This is the name of your package
version = "0.6.0-dev"
description = "Add your description here"
authors = [
    {name = "QuEra Computing Inc.", email = "info@quera.com"},
]
dependencies = [
    "bloqade-geometry >=0.1.2,<0.2.0-dev",
    "bloqade-circuit ~= 0.4.0",
    "kirin-toolchain ~= 0.17.0",
    "pint>=0.24.3",
    "pyarrow>=17.0.0",
    "pytest>=8.3.2",
    "numpy>=1.26.4",
]

readme = "README.md"
requires-python = ">= 3.10"

[project.optional-dependencies]
visualization = [
    "matplotlib>=3.9.4",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "coverage>=7.6.4",
    "ipython>=8.29.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "numpy>=1.26.4",
    "pre-commit>=4.0.1",
    "pyright[nodejs]>=1.1.388",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
    "rust-just>=1.36.0",
]

[dependency-groups]
doc = [
    "griffe-inherited-docstrings>=1.1.1",
    "mike>=2.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-material>=9.5.44",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocstrings[python]>=0.27.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bloqade"]

[tool.isort]
profile = "black"
combine_as_imports = true
multi_line_output = 3
src_paths = ["src/bloqade"]

[tool.black]
line-length = 88

[tool.ruff]
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.coverage.run]
include = ["src/bloqade/*"]

[tool.pytest.ini_options]
testpaths = "test/"
